fix: configure env values when creating a feature
Previously this would create the feature, but not set any values.  And
since `persistFeatures` loops over the deployed features, not the
configured ones, it was not configured there either.

Since it could be possible that we would want to create a feature, and
not change any others, it makes sense to configure it at creation time.

Fixes #14